Restructured package for more logical structure
