+++ .gitignore
@@ -2,3 +2,4 @@
 dist/*
 src/*
 !src/pybalmorel
+tests/files/DE.inc
