+++ src/pybalmorel/__init__.py
@@ -1,1 +1,3 @@
 from . import functions, formatting
+
+from .classes import IncFile, MainResults
